refactor for parsing unstructured log events
refactor the lambda to reduce code duplication and allowing parsing of
unstructured log events.

not all logs processed by the lambda are structured as parsable JSON
documents (we see a large number of logs originating from the
control-plane that are unstructured), which results in the lambda
failing and ultimately dropping logs,

to address this we break the code up into smaller reusable functions and
attempt to process log events as "raw events" in circumstances where
they are either unrecognised or fail to parse as json.

In addition to this, the redundant VPC flow log code paths have been
dropped.